Separate patricia tests
