+++ test/colony-network-mining.js
@@ -2,7 +2,6 @@
 
 import path from "path";
 import BN from "bn.js";
-import web3Utils from "web3-utils";
 import { TruffleLoader } from "@colony/colony-js-contract-loader-fs";
 
 import { forwardTime, checkErrorRevert, web3GetTransactionReceipt } from "../helpers/test-helper";
@@ -1696,114 +1695,5 @@ contract("ColonyNetworkStaking", accounts => {
 
     it.skip("The reputation mining client should calculate reputation decay correctly");
     it.skip("should abort if a deposit did not complete correctly");
-  });
-
-  describe("Javascript Patricia Tree implementation", () => {
-    it("should have identical root hashes after one insert", async () => {
-      const jsClient = new ReputationMiner({
-        loader: contractLoader,
-        minerAddress: MAIN_ACCOUNT,
-        realProviderPort: REAL_PROVIDER_PORT,
-        useJsTree: true
-      });
-      const solClient = new ReputationMiner({
-        loader: contractLoader,
-        minerAddress: MAIN_ACCOUNT,
-        realProviderPort: REAL_PROVIDER_PORT,
-        useJsTree: false
-      });
-
-      await jsClient.initialise(colonyNetwork.address);
-      await solClient.initialise(colonyNetwork.address);
-
-      const dog = web3Utils.fromAscii("dog");
-      const fido = web3Utils.fromAscii("fido");
-
-      await jsClient.reputationTree.insert(dog, fido);
-      await solClient.reputationTree.insert(dog, fido);
-
-      const jsRoot = await jsClient.reputationTree.getRootHash();
-      const solRoot = await solClient.reputationTree.getRootHash();
-      assert.equal(jsRoot, solRoot);
-    });
-
-    it("should have identical root hashes after two inserts and one update", async () => {
-      const jsClient = new ReputationMiner({
-        loader: contractLoader,
-        minerAddress: MAIN_ACCOUNT,
-        realProviderPort: REAL_PROVIDER_PORT,
-        useJsTree: true
-      });
-      const solClient = new ReputationMiner({
-        loader: contractLoader,
-        minerAddress: MAIN_ACCOUNT,
-        realProviderPort: REAL_PROVIDER_PORT,
-        useJsTree: false
-      });
-
-      await jsClient.initialise(colonyNetwork.address);
-      await solClient.initialise(colonyNetwork.address);
-
-      const dog = web3Utils.fromAscii("dog");
-      const fido = web3Utils.fromAscii("fido");
-      const ape = web3Utils.fromAscii("ape");
-      const bubbles = web3Utils.fromAscii("bubbles");
-      const rover = web3Utils.fromAscii("rover");
-
-      await jsClient.reputationTree.insert(dog, fido);
-      await solClient.reputationTree.insert(dog, fido);
-
-      await jsClient.reputationTree.insert(ape, bubbles);
-      await solClient.reputationTree.insert(ape, bubbles);
-
-      await jsClient.reputationTree.insert(dog, rover);
-      await solClient.reputationTree.insert(dog, rover);
-
-      const jsRoot = await jsClient.reputationTree.getRootHash();
-      const solRoot = await solClient.reputationTree.getRootHash();
-      assert.equal(jsRoot, solRoot);
-    });
-
-    it("should give identical proofs after two inserts and one update", async () => {
-      const jsClient = new ReputationMiner({
-        loader: contractLoader,
-        minerAddress: MAIN_ACCOUNT,
-        realProviderPort: REAL_PROVIDER_PORT,
-        useJsTree: true
-      });
-      const solClient = new ReputationMiner({
-        loader: contractLoader,
-        minerAddress: MAIN_ACCOUNT,
-        realProviderPort: REAL_PROVIDER_PORT,
-        useJsTree: false
-      });
-
-      await jsClient.initialise(colonyNetwork.address);
-      await solClient.initialise(colonyNetwork.address);
-
-      const dog = web3Utils.fromAscii("dog");
-      const fido = web3Utils.fromAscii("fido");
-      const ape = web3Utils.fromAscii("ape");
-      const bubbles = web3Utils.fromAscii("bubbles");
-      const rover = web3Utils.fromAscii("rover");
-
-      await jsClient.reputationTree.insert(dog, fido);
-      await solClient.reputationTree.insert(dog, fido);
-
-      await jsClient.reputationTree.insert(ape, bubbles);
-      await solClient.reputationTree.insert(ape, bubbles);
-
-      await jsClient.reputationTree.insert(dog, rover);
-      await solClient.reputationTree.insert(dog, rover);
-
-      const [jsMask, jsSiblings] = await jsClient.reputationTree.getProof(dog);
-      const [solMask, solSiblings] = await solClient.reputationTree.getProof(dog);
-
-      assert.equal(jsMask.toString(), solMask.toString());
-      assert.equal(jsSiblings.length, solSiblings.length);
-      for (let i = 0; i < jsSiblings.length; i += 1) {
-        assert.equal(jsSiblings[i], solSiblings[i]);
-      }
-    });
   });
 });
